Remove IssueTrackerType.all_issues_link()
this was used in TestRun Report page to show a single link
that will open all bugs in the Issue Tracker. Most trackers don't
support this and the UI portion has been replaced so we don't
use this anymore.

NOTE:
Now that we use URLs to link to bugs this could probably be brought
back by trying to filter off the common parts of all URLs and
leave only the bug IDs (string).

TODO: how would that work if we have multiple issues, from multiple
issue trackers reported in the same TestRun ? How do we know which
issues belong to which issue tracker? What if one of the trackers
doesn't support URLs to show all issues? Do we just display incomplete
information or not display anything or display warnings ?

Or just forget about this and display a list of all possible bugs
linked to this TestRun so the user can navigate them. Maybe add
a JavaScript button that will open all URLs into separate tabs
for their convenience.